+++ crates/client/sync/src/fetch/mod.rs
@@ -7,11 +7,13 @@ use mc_db::MadaraBackend;
 use mc_gateway_client::GatewayProvider;
 use mc_rpc::versions::admin::v0_1_0::MadaraStatusRpcApiV0_1_0Client;
 use mp_gateway::error::{SequencerError, StarknetError, StarknetErrorCode};
-use mp_utils::{channel_wait_or_graceful_shutdown, service::ServiceContext, wait_or_graceful_shutdown};
+use mp_utils::service::ServiceContext;
 use tokio::sync::{mpsc, oneshot};
 use url::Url;
 
 use crate::fetch::fetchers::fetch_block_and_updates;
+
+use self::fetchers::WarpUpdateConfig;
 
 pub mod fetchers;
 
@@ -23,23 +25,27 @@ pub struct L2FetchConfig {
     pub n_blocks_to_sync: Option<u64>,
     pub stop_on_sync: bool,
     pub sync_parallelism: usize,
-    pub warp_update: bool,
-    pub warp_update_port_rpc: u16,
-    pub warp_update_port_fgw: u16,
+    pub warp_update: Option<WarpUpdateConfig>,
 }
 
 pub async fn l2_fetch_task(
     backend: Arc<MadaraBackend>,
     provider: Arc<GatewayProvider>,
-    ctx: ServiceContext,
+    mut ctx: ServiceContext,
     mut config: L2FetchConfig,
 ) -> anyhow::Result<()> {
     // First, catch up with the chain
-    // let backend = &backend;
+    let L2FetchConfig { first_block, ref warp_update, .. } = config;
 
-    let L2FetchConfig { first_block, warp_update, warp_update_port_rpc, warp_update_port_fgw, .. } = config;
-
-    if warp_update {
+    if let Some(WarpUpdateConfig {
+        warp_update_port_rpc,
+        warp_update_port_fgw,
+        warp_update_shutdown_sender,
+        warp_update_shutdown_receiver,
+        deferred_service_start,
+        deferred_service_stop,
+    }) = warp_update
+    {
         let client = jsonrpsee::http_client::HttpClientBuilder::default()
             .build(format!("http://localhost:{warp_update_port_rpc}"))
             .expect("Building client");
@@ -62,15 +68,29 @@ pub async fn l2_fetch_task(
             .unwrap_or(NonZeroUsize::new(1usize).expect("1 should always be in usize bound"));
         config.sync_parallelism = Into::<usize>::into(available_parallelism) * 2;
 
-        let next_block = match sync_blocks(backend.as_ref(), &provider, &ctx, &config).await? {
+        let next_block = match sync_blocks(backend.as_ref(), &provider, &mut ctx, &config).await? {
             SyncStatus::Full(next_block) => next_block,
             SyncStatus::UpTo(next_block) => next_block,
         };
 
-        if client.shutdown().await.is_err() {
-            tracing::error!("❗ Failed to shutdown warp update sender");
-            ctx.cancel_global();
-            return Ok(());
+        if *warp_update_shutdown_sender {
+            if client.shutdown().await.is_err() {
+                tracing::error!("❗ Failed to shutdown warp update sender");
+                ctx.cancel_global();
+                return Ok(());
+            }
+
+            for svc_id in deferred_service_stop {
+                ctx.service_remove(*svc_id);
+            }
+
+            for svc_id in deferred_service_start {
+                ctx.service_add(*svc_id);
+            }
+        }
+
+        if *warp_update_shutdown_receiver {
+            return anyhow::Ok(());
         }
 
         config.n_blocks_to_sync = config.n_blocks_to_sync.map(|n| n - (next_block - first_block));
@@ -78,7 +98,7 @@ pub async fn l2_fetch_task(
         config.sync_parallelism = save;
     }
 
-    let mut next_block = match sync_blocks(backend.as_ref(), &provider, &ctx, &config).await? {
+    let mut next_block = match sync_blocks(backend.as_ref(), &provider, &mut ctx, &config).await? {
         SyncStatus::Full(next_block) => {
             tracing::info!("🥳 The sync process has caught up with the tip of the chain");
             next_block
@@ -105,17 +125,27 @@ pub async fn l2_fetch_task(
 
         let mut interval = tokio::time::interval(sync_polling_interval);
         interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-        while wait_or_graceful_shutdown(interval.tick(), &ctx).await.is_some() {
-            loop {
-                match fetch_block_and_updates(&backend.chain_config().chain_id, next_block, &provider, &ctx).await {
+        while ctx.run_until_cancelled(interval.tick()).await.is_some() {
+            // It is possible the chain produces multiple blocks in the span of
+            // a single loop iteration, so we keep fetching until we reach the
+            // tip again.
+            let chain_id = &backend.chain_config().chain_id;
+            let fetch = |next_block: u64| fetch_block_and_updates(chain_id, next_block, &provider);
+
+            while let Some(block) = ctx.run_until_cancelled(fetch(next_block)).await {
+                match block {
                     Err(FetchError::Sequencer(SequencerError::StarknetError(StarknetError {
                         code: StarknetErrorCode::BlockNotFound,
                         ..
                     }))) => {
                         break;
                     }
-                    val => {
-                        if fetch_stream_sender.send(val?).await.is_err() {
+                    Err(e) => {
+                        tracing::debug!("Failed to poll latest block: {e}");
+                        return Err(e.into());
+                    }
+                    Ok(unverified_block) => {
+                        if fetch_stream_sender.send(unverified_block).await.is_err() {
                             // stream closed
                             break;
                         }
@@ -126,7 +156,8 @@ pub async fn l2_fetch_task(
             }
         }
     }
-    Ok(())
+
+    anyhow::Ok(())
 }
 
 /// Whether a chain has been caught up to the tip or only a certain block number
@@ -154,27 +185,24 @@ enum SyncStatus {
 async fn sync_blocks(
     backend: &MadaraBackend,
     provider: &Arc<GatewayProvider>,
-    ctx: &ServiceContext,
+    ctx: &mut ServiceContext,
     config: &L2FetchConfig,
 ) -> anyhow::Result<SyncStatus> {
     let L2FetchConfig { first_block, fetch_stream_sender, n_blocks_to_sync, sync_parallelism, .. } = config;
 
     // Fetch blocks and updates in parallel one time before looping
-    let fetch_stream =
-        (*first_block..).take(n_blocks_to_sync.unwrap_or(u64::MAX) as _).map(|block_n| {
-            let provider = Arc::clone(provider);
-            let ctx = ctx.clone();
-            async move {
-                (block_n, fetch_block_and_updates(&backend.chain_config().chain_id, block_n, &provider, &ctx).await)
-            }
-        });
+    let fetch_stream = (*first_block..).take(n_blocks_to_sync.unwrap_or(u64::MAX) as _).map(|block_n| {
+        let provider = Arc::clone(provider);
+        let chain_id = &backend.chain_config().chain_id;
+        async move { (block_n, fetch_block_and_updates(chain_id, block_n, &provider).await) }
+    });
 
     // Have `sync_parallelism` fetches in parallel at once, using futures Buffered
     let mut next_block = *first_block;
     let mut fetch_stream = stream::iter(fetch_stream).buffered(*sync_parallelism);
 
-    loop {
-        let Some((block_n, val)) = channel_wait_or_graceful_shutdown(fetch_stream.next(), ctx).await else {
+    while let Some(next) = ctx.run_until_cancelled(fetch_stream.next()).await {
+        let Some((block_n, val)) = next else {
             return anyhow::Ok(SyncStatus::UpTo(next_block));
         };
 
@@ -195,6 +223,8 @@ async fn sync_blocks(
 
         next_block = block_n + 1;
     }
+
+    anyhow::Ok(SyncStatus::UpTo(next_block))
 }
 
 #[derive(thiserror::Error, Debug)]
@@ -253,9 +283,7 @@ mod test_l2_fetch_task {
                             n_blocks_to_sync: Some(5),
                             stop_on_sync: false,
                             sync_parallelism: 10,
-                            warp_update: false,
-                            warp_update_port_rpc: 9943,
-                            warp_update_port_fgw: 8080,
+                            warp_update: None,
                         },
                     ),
                 )